clinic/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2020.4
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
uuid==1.30
